Added PrivateImplementationPattern
Added class PrivateImplementationPattern that facilitates using the
pattern of same name by just inheriting from this class template.

There are two variants. The Shared pattern is to be used when the
instances of a class shall be able to share one Implementation instance.
The Single pattern is to be used when all instances of a class shall
each have a dedicated instance of Implementation.

Signed-off-by: Danny van Dyk <dyk@honei.org>