Made the number of pooled threads constant
The amount of parallelism shall (at first) be determined by the number
of created and dispatched tasks, not by the choice of a number of
threads. Therefore the pool will now always allocate at least as much
threads as processing units are available. Only if there is a higher
number specified in honeirc, more threads are created.

Signed-off-by: Sven Mallach <mallach@honei.org>